Fixed broken test run Groovy system.
Having Groovy on the build path brought in an
extra version of Groovy that wasn't helpful, and
which caused errors on Spock test runs in Eclipse.
The solution is to remove Groovy from the build
path in Eclipse.

Example error:

org.spockframework.util.InternalSpockError:
Failed to instantiate spec 'LambdaRequestHandlerSpec'

Caused by: groovy.lang.GroovyRuntimeException:
Conflicting module versions. Module [groovy-all is
loaded in version 2.4.10 and you are trying to load
version 2.4.12